+++ Module/en-US/SplitPipeline.dll-Help.ps1
@@ -15,16 +15,17 @@ Import-Module SplitPipeline
 '@
 	description = @'
     The cmdlet splits the input, processes its parts by parallel pipelines, and
-    outputs the results for further processing. It works without collecting the
-    entire input, large or even infinite.
+    outputs the results for further processing. It may work without collecting
+    the whole input, large or infinite.
 
-	If processing is relatively fast then it is important to specify the Load
-	limits. Otherwise Split-Pipeline may work slower than a standard pipeline.
+	When Load is omitted the whole input is collected and split evenly between
+	Count parallel pipelines. This method shows the best performance in simple
+	cases. In other cases, e.g. on large or slow input, Load should be used in
+	order to enable processing of partially collected input.
 
 	The cmdlet creates several pipelines. Each pipeline is created when input
 	parts are available, created pipelines are busy, and their number is less
-	than Count. Each created pipeline is used for processing several input
-	parts, one at a time.
+	than Count. Each pipeline is used for processing one or more input parts.
 
 	Because each pipeline works in its own runspace variables, functions, and
 	modules from the main script are not automatically available for pipeline
@@ -99,24 +100,26 @@ Import-Module SplitPipeline
 		much processor resources increasing the number may improve performance.
 '@
 		Load = @'
-		One or two values specifying the minimum and maximum number of input
-		objects for each parallel pipeline. The first value is the recommended
-		minimum, 1 is the default. The second value is the maximum, not limited
-		by default.
+		Enables processing of partially collected input and specifies input
+		part limits. If it is omitted then the whole input is collected and
+		split evenly between pipelines.
+
+		The parameter accepts an array of one or two integers. The first is the
+		minimum number of objects per pipeline. The second is the maximum, not
+		limited if omitted.
 
 		If processing is fast then it is important to specify a proper minimum.
 		Otherwise Split-Pipeline may work even slower than a standard pipeline.
 
-		Setting the maximum number causes more frequent output if the limit is
-		actually hit. This may be important for feeding downstream commands in
-		the pipeline working at the same time.
+		Setting the maximum causes more frequent output. For example, this may
+		be important for feeding simultaneously working downstream pipelines.
 
 		Setting the maximum number is also needed for potentially large input
 		in order to limit the input queue size and avoid out of memory issues.
 		The maximum queue size is set internally to Load[1] * Count.
 
-		Use the switch Verbose in order to get statistics messages with data
-		which may help to choose reasonable load limits.
+		Use the switch Verbose in order to get some statistics which may help
+		to choose suitable load limits.
 
 		CAUTION: The queue limit may be ignored and exceeded if Refill is used.
 		Any number of objects written via [ref] go straight to the input queue.
@@ -165,8 +168,8 @@ Import-Module SplitPipeline
 		@{
 			type = 'Object'
 			description = @'
-		Output of the Begin, Script, and End script blocks. The Begin and End
-		scripts are invoked once for each pipeline before and after processing.
+		Output of the Begin, Script, and End script blocks. The scripts Begin
+		and End are invoked once for each pipeline before and after processing.
 		The script Script is invoked repeatedly with input parts piped to it.
 '@
 		}
@@ -174,8 +177,8 @@ Import-Module SplitPipeline
 	examples = @(
 		@{
 			code = {
-				1..10 | . {process{$_; sleep 1}}
-				1..10 | Split-Pipeline -Count 10 {process{$_; sleep 1}}
+				1..10 | . {process{ $_; sleep 1 }}
+				1..10 | Split-Pipeline -Count 10 {process{ $_; sleep 1 }}
 			}
 			remarks = @'
 	Two commands perform the same job simulating long but not processor
@@ -198,8 +201,8 @@ Import-Module SplitPipeline
 				}} | Measure-Object -Sum
 			}
 			remarks = @'
-	This is an example of Split-Pipeline with refilled input. The convention:
-	[ref] objects refill the input, other objects go to output as usual.
+	This is an example of Split-Pipeline with refilled input. By the convention
+	output [ref] objects refill the input, other objects go to output as usual.
 
 	The code calculates the number and size of files in $PSHOME. It is a "how
 	to" sample, performance gain is not expected because the code is trivial
@@ -217,16 +220,16 @@ Import-Module SplitPipeline
 	scripts. Such items should be specified by Variable, Function, and Module
 	parameters in order to be available.
 
-    > $arr = @('one', 'two', 'three'); 0..2 | ForEach-Object {$arr[$_]}
+    > $arr = @('one', 'two', 'three'); 0..2 | . {process{ $arr[$_] }}
     one
     two
     three
 
-    > $arr = @('one', 'two', 'three'); 0..2 | Split-Pipeline {process{$arr[$_]}}
+    > $arr = @('one', 'two', 'three'); 0..2 | Split-Pipeline {process{ $arr[$_] }}
     Split-Pipeline : Cannot index into a null array.
     ...
 
-    > $arr = @('one', 'two', 'three'); 0..2 | Split-Pipeline -Variable arr {process {$arr[$_]}}
+    > $arr = @('one', 'two', 'three'); 0..2 | Split-Pipeline -Variable arr {process{ $arr[$_] }}
     one
     two
     three