+++ Src/AssemblyInfo.cs
@@ -4,10 +4,10 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyProduct("SplitPipeline")]
-[assembly: AssemblyVersion("1.3.1")]
-[assembly: AssemblyTitle("SplitPipeline - Parallel Data Processing in PowerShell")]
-[assembly: AssemblyCompany("https://github.com/nightroman/SplitPipeline")]
+[assembly: AssemblyVersion("1.4.0")]
 [assembly: AssemblyCopyright("Copyright (c) 2011-2014 Roman Kuzmin")]
+[assembly: AssemblyCompany("https://github.com/nightroman/SplitPipeline")]
+[assembly: AssemblyTitle("SplitPipeline - Parallel Data Processing in PowerShell")]
 
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]